perf(viewer): CHOLESKY semantic → Float16 on GPU (GSplats)
Phase-4 C-ts-4 (final commit): activate Float16 GPU dtype for the
CHOLESKY semantic on GSplats. The shader reads cholesky as `float`
and runs the covariance inversion in single precision, so the
quantization is bounded to the storage step. Disk format
unchanged; narrowing happens in the packer at upload time.

Map flip: `DEFAULT_GPU_DTYPE_BY_SEMANTIC.cholesky` flipped from
`'float32'` to `'float16'`.

Precision analysis:
- Float16 mantissa: 11 bits → ~5e-4 relative precision.
- Subnormal floor: ~6e-5. Splats with cholesky entries below this
  underflow; in practice this means aspect ratios > 16000:1 will
  see the smallest axis quantize to zero.
- Inverse-covariance computation (Mahalanobis distance in the
  fragment shader) runs in single precision after the GPU widens
  the Float16 attribute. Only the STORAGE is quantized; the math
  isn't.
- Typical scientific scene splats: < 100:1 aspect ratio. Float16
  cholesky storage is well within parity-spec tolerance for these.

Escape hatch: if a future workload sees parity regression on
extreme anisotropy, override per-attribute via:

  { name: 'aCholesky01', ..., semantic: 'cholesky', dtype: 'float32' }

This keeps cholesky at Float32 while everything else stays
narrowed. ~12 bytes/splat lost back from C-ts-4 — modest cost
for safety on a workload that needs it.

Bandwidth impact: 3 × 2 × 4 = 24 bytes/splat → 3 × 2 × 2 = 12
bytes/splat. 12 bytes saved per splat on the cholesky alone.

Cumulative phase-4 savings on GSplats (C-ts-2 + C-ts-4):
- aColor: 6 bytes/inst saved
- aAmplitude: 2 bytes/inst saved (C-ts-3 POSITIVE_SCALAR)
- aCholesky01/23/45: 12 bytes/inst saved (C-ts-4)
Total: 20 bytes/splat saved.

Gates:
- pnpm typecheck: clean
- pnpm test --run: 4740 passed
- tsl-shader-parity: 13/13 pass — Float16 cholesky storage does
  NOT regress GSplats parity at the parity-spec's mean-abs-diff
  threshold. Typical splats render identically through both
  backends.
- tsl-codegen-snapshot: zero diff
- E2E smoke (geometry-types): 5/5 pass — GSplats mesh attributes
  bind and render correctly with Float16 cholesky.

This closes phase-4 GPU-side narrowing. BOUNDED_SCALAR (sharpness,
clipped) stays at Float32 — Uint8-norm narrowing for it is
deferred (C-ts-3b future) pending encoder-bounds metadata
threading.

+++ packages/luxar-viewer/src/rendering/interleaved-attributes.ts
@@ -189,7 +189,15 @@ const DEFAULT_GPU_DTYPE_BY_SEMANTIC: Record<SemanticType, GpuDtype> = {
   // `radiusScale` uniforms — non-trivial because bounds vary per
   // dataset. Stays Float32 for now.
   bounded_scalar: 'float32',
-  cholesky: 'float32',
+  // C-ts-4: CHOLESKY (GSplats covariance) narrowed to Float16 on
+  // GPU. Float16 mantissa (~5e-4 relative precision) is adequate
+  // for typical splat aspect ratios up to ~1000:1. The shader
+  // reads cholesky as `float` and runs inversion in single
+  // precision, so the quantization is bounded to the storage step.
+  // Per-attribute override (`dtype: 'float32'` on specific
+  // gsplat specs) is available if parity regresses on extreme
+  // anisotropy.
+  cholesky: 'float16',
   // INDEX attributes don't flow through this packer (they're handled
   // by the geometry's element-array buffer); the value is here for
   // type-coverage only.
